# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/rates/plans": {
      "get": {
        "operationId": "RatesController_listPlans",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rates"
        ]
      },
      "post": {
        "operationId": "RatesController_createPlan",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Rates"
        ]
      }
    },
    "/api/v1/rates/calendar": {
      "post": {
        "operationId": "RatesController_setDailyRates",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Rates"
        ]
      },
      "get": {
        "operationId": "RatesController_calendar",
        "parameters": [
          {
            "name": "ratePlanId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rates"
        ]
      }
    },
    "/api/v1/rates/quote": {
      "get": {
        "operationId": "RatesController_quote",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ratePlanId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "arrival",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departure",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rates"
        ]
      }
    },
    "/api/v1/properties/tax-rules": {
      "get": {
        "operationId": "RatesController_listTaxRules",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rates"
        ]
      },
      "post": {
        "operationId": "RatesController_upsertTaxRule",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Rates"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lodgiva API",
    "description": "Hotel management platform API. All protected routes require a Bearer access token; the tenant is always derived from the token and never from the request body.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}